% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.r
\docType{methods}
\name{bind}
\alias{bind}
\alias{c.madness}
\alias{cbind2,madness,missing-method}
\alias{cbind2,madness,madness-method}
\alias{cbind2,madness,ANY-method}
\alias{cbind2,ANY,madness-method}
\alias{rbind2,madness,missing-method}
\alias{rbind2,madness,madness-method}
\alias{rbind2,madness,ANY-method}
\alias{rbind2,ANY,madness-method}
\title{Row and Column Bind}
\usage{
\\method{c}{madness}(...)

\S4method{cbind2}{madness,missing}(x, y, ...)

\S4method{cbind2}{madness,madness}(x, y, ...)

\S4method{cbind2}{madness,ANY}(x, y, ...)

\S4method{cbind2}{ANY,madness}(x, y, ...)

\S4method{rbind2}{madness,missing}(x, y, ...)

\S4method{rbind2}{madness,madness}(x, y, ...)

\S4method{rbind2}{madness,ANY}(x, y, ...)

\S4method{rbind2}{ANY,madness}(x, y, ...)
}
\arguments{
\item{...}{optional arguments for methods (ignored here).}

\item{x, y}{\code{madness} or array, numeric, matrix objects.}
}
\description{
Row and Column Bind
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
