% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elwise.r
\docType{methods}
\name{elwise}
\alias{elwise}
\alias{abs,madness-method}
\alias{exp,madness-method}
\alias{log,madness-method}
\alias{log10,madness-method}
\alias{sqrt,madness-method}
\alias{sin,madness-method}
\alias{cos,madness-method}
\alias{tan,madness-method}
\title{Element-wise Multivariate Operations}
\usage{
\S4method{abs}{madness}(x)

\S4method{exp}{madness}(x)

\S4method{log}{madness}(x)

\S4method{log10}{madness}(x)

\S4method{sqrt}{madness}(x)

\S4method{sin}{madness}(x)

\S4method{cos}{madness}(x)

\S4method{tan}{madness}(x)
}
\arguments{
\item{x}{\code{madness} object.}
}
\description{
Element-wise multivariate operations.
}
\details{
These operations are scalar-to-scalar operations applied to each element of a 
multidimensional array.
}
\note{
The \code{exp}, \code{log}, and \code{sqrt}
functions are not to be confused with the matrix-wise operations,
\code{expm}, \code{logm} and \code{sqrtm}
}
\seealso{
\link{matwise}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
