% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madness_pkg.r
\docType{package}
\name{madness-pkg}
\alias{madness-pkg}
\alias{madness-pkg-package}
\title{Multivariate Automatic Differentiation.}
\description{
Automatic Differentiation of Matrix Operations.
}
\note{
This package is maintained as a hobby.
}
\section{Legal Mumbo Jumbo}{


madness is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU Lesser General Public License for more details.
}

\references{
Griewank, Andreas and Walther, Andrea. "Evaluating Derivatives: principles and techniques of algorithmic differentiation."
SIAM (2008).

Petersen, Kaare Brandt and Pedersen, Michael Syskind. "The Matrix Cookbook."
Technical University of Denmark (2012). 
\url{http://www2.imm.dtu.dk/pubdb/p.php?3274}

Magnus, Jan R. and Neudecker, H. "Matrix Differential Calculus with Applications in Statistics and Econometrics."
3rd Edition. Wiley Series in Probability and Statistics: Texts and References Section (2007).

Magnus, Jan R. and Neudecker, H. "The elimination matrix: some lemmas and applications," 
SIAM Journal on Algebraic Discrete Methods 1, no. 4 (1980): 422-449.
\url{http://www.janmagnus.nl/papers/JRM008.pdf}

Magnus, Jan R. and Neudecker, H. "Symmetry, 0-1 Matrices and Jacobians,"
Econometric Theory 2 (1986): 157-190.
\url{http://www.janmagnus.nl/papers/JRM014.pdf},

Fackler, Paul L. "Notes on Matrix Calculus." (2005).
\url{http://www4.ncsu.edu/~pfackler/MatCalc.pdf}

MM: DO use  @importFrom (whimps use full imports ..) !
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}

Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{package}
