% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.r
\name{matrix.trace}
\alias{matrix.trace}
\alias{matrix.trace,ANY-method}
\alias{matrix.trace,matrix-method}
\alias{matrix.trace,madness-method}
\title{Matrix Trace}
\usage{
matrix.trace(x)

\S4method{matrix.trace}{ANY}(x)

\S4method{matrix.trace}{madness}(x)
}
\arguments{
\item{x}{\code{madness} object.}
}
\description{
Matrix Trace
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
