% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madness_pkg.r
\name{madness-NEWS}
\alias{madness-NEWS}
\title{News for package \sQuote{madness}:}
\description{
News for package \sQuote{madness}.

\newcommand{\pkg}{#1}
\newcommand{\CRANpkg}{\href{https://CRAN.R-project.org/package=#1}{\pkg{#1}}}
\newcommand{\madness}{\CRANpkg{madness}}
}
\section{\madness{} Version 0.2.3 (2018-02-14) }{

\itemize{
\item emergency CRAN release to deal with failing tests
under alternative BLAS/LAPACK libraries.
}
}

\section{\madness{} Version 0.2.2 (2017-04-26) }{

\itemize{
\item emergency CRAN release for upstream changes to \code{diag}.
thanks to Martin Maechler for the patch.
}
}

\section{\madness{} Version 0.2.1 (2017-04-13) }{

\itemize{
\item emergency CRAN release for failed build.
\item no new functionality.
}
}

\section{\madness{} Version 0.2.0 (2016-01-19) }{

\itemize{
\item add static vignette.
\item modify \code{twomoments}.
\item release to CRAN.
}
}

\section{\madness{} Version 0.1.0.400 (2016-01-12) }{

\itemize{
\item adding \code{max} and \code{min}.
}
}

\section{\madness{} Version 0.1.0.300 (2016-01-10) }{

\itemize{
\item adding \code{eigen}.
}
}

\section{\madness{} Version 0.1.0.200 (2016-01-07) }{

\itemize{
\item exporting \code{diag}.
}
}

\section{\madness{} Version 0.1.0 (2015-12-15) }{

\itemize{
\item first CRAN release.
}
}

\section{\madness{} Initial Version 0.0.0.5000 (2015-12-01) }{

\itemize{
\item first github release.
}
}

