% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.r
\docType{methods}
\name{norm}
\alias{norm}
\alias{maxeig}
\alias{maxeig,madness-method}
\alias{norm,madness-method}
\alias{norm,madness,missing-method}
\alias{norm,madness,ANY-method}
\title{Matrix and vector norms.}
\usage{
maxeig(x)

\S4method{maxeig}{madness}(x)

\S4method{norm}{madness,missing}(x)

\S4method{norm}{madness,ANY}(x, type = "One")
}
\arguments{
\item{x}{\code{madness} object.}

\item{type}{character string, specifying the \emph{type} of matrix
    norm to be computed.
    A character indicating the type of norm desired.
    \describe{
      \item{\code{"O"}, \code{"o"} or \code{"1"}}{specifies the \bold{o}ne norm,
	(maximum absolute column sum);}
      \item{\code{"I"} or \code{"i"}}{specifies the \bold{i}nfinity norm (maximum
	absolute row sum);}
      \item{\code{"F"} or \code{"f"}}{specifies the \bold{F}robenius norm (the
	Euclidean norm of \code{x} treated as if it were a vector);}
      \item{\code{"M"} or \code{"m"}}{specifies the \bold{m}aximum modulus of
	all the elements in \code{x}; and}
      \item{\code{"2"}}{specifies the \dQuote{spectral} or 2-norm, which
	is the largest singular value (\code{\link{svd}}) of \code{x}.}
    }
    The default is \code{"O"}.  Only the first character of
    \code{type[1]} is used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the matrix norm, a non-negative number.
}
\description{
Compute the norm of a vector or matrix, as determined by the
\code{type}.
}
\note{
This should probably be fixed to return a scalar, not a 1 by 1 matrix?
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
