% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.r
\docType{methods}
\name{setter}
\alias{setter}
\alias{varx<-}
\alias{varx<-,madness-method}
\alias{vtag<-}
\alias{vtag<-,madness-method}
\alias{xtag<-}
\alias{xtag<-,madness-method}
\title{Setter methods.}
\usage{
xtag(x) <- value

\S4method{xtag}{madness}(x) <- value

vtag(x) <- value

\S4method{vtag}{madness}(x) <- value

varx(x) <- value

\S4method{varx}{madness}(x) <- value
}
\arguments{
\item{x}{a \code{madness} object.}

\item{value}{the new value of the tag or derivative.}
}
\description{
Modify slot data of a \code{madness} object. Note that the value and the
derivative cannot easily be changed, as allowing this form of access would
likely result in badly computed derivatives.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}

