% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.r
\docType{methods}
\name{show}
\alias{show}
\alias{show,madness-method}
\title{show a madness object.}
\usage{
show(object)

\S4method{show}{madness}(object)
}
\arguments{
\item{object}{a \code{madness} object.}
}
\description{
Displays the madness object.
}
\examples{
obj <- madness(val=matrix(rnorm(10*10),nrow=10), xtag="foo", ytag="foo^2")
obj
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}

