% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.r
\docType{methods}
\name{solve}
\alias{solve}
\alias{solve,ANY,madness-method}
\alias{solve,ANY,missing-method}
\alias{solve,array,madness-method}
\alias{solve,madness,ANY-method}
\alias{solve,madness,array-method}
\alias{solve,madness,madness-method}
\alias{solve,madness,missing-method}
\title{Basic Matrix Inversion}
\usage{
\S4method{solve}{ANY,missing}(a, b)

\S4method{solve}{madness,missing}(a, b)

\S4method{solve}{madness,madness}(a, b)

\S4method{solve}{madness,array}(a, b)

\S4method{solve}{madness,ANY}(a, b)

\S4method{solve}{array,madness}(a, b)

\S4method{solve}{ANY,madness}(a, b)
}
\arguments{
\item{a, b}{\code{madness} object or matrix value.}
}
\description{
Basic Matrix Inversion
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}

