% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.r
\docType{methods}
\name{todiag}
\alias{diag}
\alias{diag,madness-method}
\alias{todiag}
\alias{todiag,madness-method}
\title{Diagonal Operations}
\usage{
diag(x, nrow, ncol)

\S4method{diag}{madness}(x, nrow, ncol)

todiag(x)

\S4method{todiag}{madness}(x)
}
\arguments{
\item{x}{\code{madness} object.}

\item{nrow, ncol}{Optional dimensions for the result when \code{x} is not a matrix value (NYI here)}
}
\description{
Diagonal Operations
}
\note{
the (somewhat odd) use of \code{stats::diag} for two different
functions is \emph{not} repeated here, at least for now.
The \code{nrow} and \code{ncol} are ignored.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\seealso{
\code{\link{reshapes}}
}

