% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.r
\docType{methods}
\name{vec}
\alias{vec}
\alias{vec,array-method}
\alias{vec,madness-method}
\alias{vech}
\alias{vech,array-method}
\alias{vech,madness-method}
\title{vectorize a multidimensional array.}
\usage{
vec(x)

\S4method{vec}{madness}(x)

\S4method{vec}{array}(x)

vech(x, k = 0)

\S4method{vech}{array}(x, k = 0)

\S4method{vech}{madness}(x, k = 0)
}
\arguments{
\item{x}{a \code{madness} object or multidimensional array or matrix.}

\item{k}{the diagonal from which to subselect.}
}
\value{
a \code{madness} object or an array, of the vectorized array
or the subselected part. For the inverse operations, promotes to a
\code{madness} of a matrix, or a matrix.
}
\description{
Turn a multidimensional array into a (column) vector. 
Turn a (typically symmetric) matrix into a (column) vector of
the lower triangular part. Or reverse these
operations.
}
\examples{
y <- matrix(rnorm(16),ncol=4)
sy <- y + t(y)
vy <- vec(sy)
vmy <- vec(madness(sy))
vhy <- vech(sy)
vmhy <- vech(madness(sy))
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\seealso{
\code{\link{reshapes}}, in particular \code{tril}.
}

