% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.r
\docType{methods}
\name{accessor}
\alias{accessor}
\alias{dim,madness-method}
\alias{dvdx}
\alias{dvdx,madness-method}
\alias{length,madness-method}
\alias{val}
\alias{val,madness-method}
\alias{varx}
\alias{varx,madness-method}
\alias{xtag}
\alias{xtag,madness-method}
\alias{ytag}
\alias{ytag,madness-method}
\title{Accessor methods.}
\usage{
val(x)

\S4method{val}{madness}(x)

\S4method{dim}{madness}(x)

\S4method{length}{madness}(x)

dvdx(x)

\S4method{dvdx}{madness}(x)

xtag(x)

\S4method{xtag}{madness}(x)

ytag(x)

\S4method{ytag}{madness}(x)

varx(x)

\S4method{varx}{madness}(x)
}
\arguments{
\item{x}{a \code{madness} object.}
}
\description{
Access slot data from a \code{madness} object.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}

