% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.r
\docType{methods}
\name{as}
\alias{as}
\alias{as.array,madness-method}
\alias{as.array}
\alias{as.matrix,madness-method}
\alias{as.matrix}
\alias{as.numeric,madness-method}
\alias{as.numeric}
\title{Coerce madness to something else}
\usage{
\S4method{as.array}{madness}(x, ...)

\S4method{as.matrix}{madness}(x, ...)

\S4method{as.numeric}{madness}(x, ...)
}
\arguments{
\item{x}{a \code{madness} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Coerce as something else
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
