% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.r
\name{as}
\alias{as}
\alias{as.array.madness}
\alias{as.matrix.madness}
\alias{as.numeric.madness}
\title{Coerce madness to something else}
\usage{
\\method{as.array}{madness}(x,...)

\\method{as.matrix}{madness}(x,...)

\\method{as.numeric}{madness}(x,...)
}
\arguments{
\item{x}{a \code{madness} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Coerce as something else
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}

