% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matwise.r
\docType{methods}
\name{matwise}
\alias{matwise}
\alias{sqrtm}
\alias{sqrtm,madness-method}
\alias{chol.madness}
\alias{chol}
\title{Matrix-wise Multivariate Operations}
\usage{
\S4method{sqrtm}{madness}(x)

\method{chol}{madness}(x, ...)
}
\arguments{
\item{x}{\code{madness} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Element-wise multivariate operations.
}
\details{
These operations are operations on matrices: compute the symmetric square root
or the Cholesky factor. In the future, the matrix exponent and logarithm may
be implemented?
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
