% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.r
\docType{methods}
\name{reshapes}
\alias{diag,madness-method}
\alias{dim<-,madness,ANY-method}
\alias{dim<-,madness-method}
\alias{reshapes}
\alias{t}
\alias{t,madness-method}
\alias{todiag}
\alias{todiag,madness-method}
\alias{tril}
\alias{tril,madness-method}
\alias{triu}
\alias{triu,madness-method}
\alias{vec}
\alias{vec,madness-method}
\alias{vech}
\alias{vech,madness-method}
\title{Basic Reshape Operations}
\usage{
t(x)

\S4method{t}{madness}(x)

vec(x)

\S4method{vec}{madness}(x)

vech(x)

\S4method{vech}{madness}(x)

\S4method{diag}{madness}(x)

tril(x, k = 0, ...)

\S4method{tril}{madness}(x, k = 0)

triu(x, k = 0, ...)

\S4method{triu}{madness}(x, k = 0)

todiag(x)

\S4method{todiag}{madness}(x)

\S4method{dim}{madness}(x) <- value
}
\arguments{
\item{x}{\code{madness} object.}

\item{k}{the index of the diagonal number from which to extract.\code{madness} object.}

\item{...}{Optional arguments used by specific methods. (None used at present.)}

\item{value}{an array of the new dimensions of the object value.}
}
\description{
Basic Reshape Operations
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}

