% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.r
\name{todiag}
\alias{todiag}
\alias{diag,madness-method}
\alias{diag}
\alias{todiag,madness-method}
\title{Diagonal Operations}
\usage{
\S4method{diag}{madness}(x)

todiag(x)

\S4method{todiag}{madness}(x)
}
\arguments{
\item{x}{\code{madness} object.}
}
\description{
Diagonal Operations
}
\note{
the (somewhat odd) use of \code{stats::diag} for two different
functions is \emph{not} repeated here, at least for now.
}
\seealso{
\code{\link{reshapes}}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
